backend/pyproject.toml
[project]
name = "odi-backend"
version = "1.0.0"
description = "ODI - Final project backend"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Igor Ochocki" }
]

dependencies = [
    # Web framework
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    
    # Database
    "sqlalchemy>=2.0.36",
    "aiosqlite==0.19.0",
    
    # Data validation
    "pydantic==2.9.0",
    "pydantic-settings==2.1.0",
    "email-validator==2.1.0",
    
    # Authentication and security
    "python-jose[cryptography]==3.3.0",
    "passlib[argon2]==1.7.4",
    "argon2-cffi==23.1.0",
    "pyotp==2.9.0",
    "qrcode[pil]==7.4.2",
    
    # Cryptography
    "cryptography==41.0.7",
    "pynacl==1.5.0",
    
    # Rate limiting
    "slowapi==0.1.9",
    
    # Email service
    "aiosmtplib==3.0.1",
    "jinja2==3.1.3",
    
    # Utilities
    "python-dotenv==1.0.0",
    "httpx==0.26.0",
    
    # CORS and middleware
    "starlette==0.35.1",
    
    # Logging
    "structlog==24.1.0",
]

[project.optional-dependencies]
dev = [
    "httpx==0.26.0",
    "mypy==1.8.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy"
]
